add current command manually; fix 'sign'-liked functions bugs

+++ User_application/ACMSim.h
@@ -81,8 +81,8 @@ extern struct GlobalWatch watch;
 
 
 /* Declaration of Utility Function */
-int isNumber(double x);
-double sign(double x);
+int isNumber(REAL x);
+REAL sign(REAL x);
 inline REAL signf(REAL x) { return (x > 0) - (x < 0); }
 double fabs(double x);
 REAL _lpf(REAL x, REAL y_tminus1, REAL time_const_inv);
